refactor(wiki-compose): apply thermo-nuclear review fixes
- Align frontend backend resolution with Google-fixed Wiki Compose model
- Extract session reducer module; slim useWikiComposeSession hook
- Consolidate auto-start via startPolicy and canRetryStart
- Deduplicate compose session locale prep on API routes
- Remove duplicate props and dead outline preview ternary
